% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{dat}
\alias{dat}
\title{IRI validation dataset}
\format{
A data frame with 835 rows and 2 variables:
\describe{
\item{\code{Gender}}{integer COLUMN_DESCRIPTION}
\item{\code{IRI_EC}}{double COLUMN_DESCRIPTION}
}
}
\usage{
dat
}
\description{
This dataset consists the two variables which were used in the study examining psychometric properties of the Interpersonal Reactivity Index.
}
\details{
DETAILS
}
\keyword{datasets}
