% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_weight_summary.R
\docType{methods}
\name{edge_weight_summary}
\alias{edge_weight_summary}
\title{Summarizing Bootstrapped Network Estimates from 'bootnet'}
\format{
An object of class \code{"tibble"}.
}
\usage{
edge_weight_summary(
  bootnet_output,
  include_sample_edge_weight = TRUE,
  include_p_values = "none"
)
}
\arguments{
\item{bootnet_output}{Output from the 'bootnet' package after bootstrapping network analysis.}

\item{include_sample_edge_weight}{Logical, whether to include sample edge weight in the summary table.}

\item{include_p_values}{String, whether to include p-values as significance stars, add p-value column, or exclude p-values entirely. Options are "stars", "exact", or "none". Default is "none".}
}
\value{
A tibble summarizing edge weights.
}
\description{
Summarizing Bootstrapped Network Estimates from 'bootnet'
}
\details{
This function provides a convenient summary of bootstrapped network estimates from the 'bootnet' package, including edge weights, their confidence intervals, and p-values alternatively. It is tailored to enhance the usability of 'bootnet' output by summarizing critical metrics in a concise format. Importantly, it only summarizes results in edges that were non-zero on sample level. Edges that had zero edge weight in a sample are filtered out.

The empirical p-values are calculated using the following formula:
\deqn{p = \frac{\sum_{i=1}^{n} \mathbf{1}\{|x_i| \geq |x_0|\}}{n}}
where \eqn{x_0} is the original estimate and \eqn{x_i} are the bootstrap estimates.
}
\examples{
  # Example testing of function with 'bootnet' output
  set.seed(746841)
  library(tidyr)
  test_dat_1 <- tibble("Group" = rbinom(100, size = 1, prob = .5),
                       "y" = ifelse(Group == 0,
                                    rnorm(100, mean = 50, sd = 10),
                                    rnorm(100, mean = 40, sd = 12)),
                       "z" = rnorm(100, mean = 41, sd = 11),
                       "w" = rnorm(100, mean = 40, sd = 10))

  estimate <- bootnet::estimateNetwork(test_dat_1, default = "EBICglasso")
  boot_results <- bootnet::bootnet(estimate, nBoots = 100, nCores = 1)

  summary <- edge_weight_summary(bootnet_output = boot_results)
  print(summary)
}
\references{
Epskamp, S., Borsboom, D., & Fried, E. I. (2018). Network analysis: An integrative approach to the structure of psychopathology. Annual review of clinical psychology, 14, 91-121.
}
\author{
Lukas Novak, \email{lukasjirinovak@gmail.com}
}
\keyword{'bootnet'}
\keyword{analysis,}
\keyword{bootstrapping,}
\keyword{network}
