% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasy.lin.reg.R
\docType{data}
\name{lasy.lin.reg}
\alias{lasy.lin.reg}
\title{Lasy linear regression function}
\format{
An object of class \code{"tibble"}
}
\usage{
lasy.lin.reg(
  independent.var,
  dependent.var,
  print.cov = FALSE,
  Z_score_independent = FALSE,
  check_multicolinearity = TRUE,
  covariates = NULL,
  data
)
}
\arguments{
\item{independent.var}{independent variable/s}

\item{dependent.var}{dependent variable/s}

\item{print.cov}{Print effect of covariates, default is FALSE}

\item{Z_score_independent}{Should independent variables be z-scored? Default is FALSE}

\item{check_multicolinearity}{Should multicolinearity assumption be checked? Default is TRUE}

\item{covariates}{covariates to be included in a model}

\item{data}{data frame or tibble object}
}
\value{
data frame
}
\description{
This function performs linear regression and print results in tibble output.
This function aims to provide the results of the regression analysis in the format, which is frequently
desired in academic journals.
}
\details{
Currently, this function does not provide model fit indicators such as AIC or BIC.
}
\examples{
# data simulation
library(dplyr)
data.PAQ = tibble(.rows = 1000)
data.PAQ <- data.PAQ \%>\%
 mutate("multiple__exper_1" = rnorm(n = nrow(data.PAQ), mean = 20, sd = 5)) \%>\%
 mutate("exper_1" = rnorm(n = nrow(data.PAQ),  mean = 20, sd = 5)) \%>\%
 mutate("exper_1" = rnorm(n = nrow(data.PAQ),  mean = 20, sd = 5),
        "exper_2" = rnorm(n = nrow(data.PAQ),  mean = 20, sd = 5),
        "last_binary_vasdl" = rnorm(n = nrow(data.PAQ), mean = 20, sd = 20),
        "last_binary_val2" = rnorm(n = nrow(data.PAQ), mean = 5, sd = 5),
        "last_binary_val3" = rnorm(n = nrow(data.PAQ), mean = 10, sd = 40),
        "last_binary_val4" = rnorm(n = nrow(data.PAQ), mean = 50, sd = 5),
        "last_binary_val5" = rnorm(n = nrow(data.PAQ), mean = 5, sd = 4),
        "last_binary_val6" = rnorm(n = nrow(data.PAQ), mean = 65, sd = 5))

lin.reg.output <- lasy.lin.reg(independent.var = c("last_binary_vasdl"),
                              covariates = c("last_binary_val6"),
                              dependent.var = c("last_binary_val5","last_binary_val4"),
                              Z_score_independent = FALSE,
                              check_multicolinearity = TRUE,
                              data = data.PAQ,
                              print.cov = FALSE)

print(lin.reg.output)
}
\references{
Welch, B. L. (1947). "The generalization of "Student's"
problem when several different population variances are involved".
Biometrika. 34 (1--2): 28--35.

Wilcoxon, F., Individual Comparisons by Ranking Methods,
Biometrics Bulletin, Vol. 1, 1945, pp. 80--83. DOI:10.2307/3001968

Dunn, O. J. (1961) Multiple comparisons among means.
Journal of the American Statistical Association. 56, 52--64.

Games, P. A., Keselman, H. J., & Clinch, J. J.
Tests for homogeneity of variance in factorial designs.
Psychological Bulletin, 86, 978--984
}
\author{
Lukas Novak, \email{lukasjirinovak@gmail.com}
}
\keyword{linear-regression}
