% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasy.log.reg.R
\docType{data}
\name{lasy.log.reg}
\alias{lasy.log.reg}
\title{Lasy logistic regression function}
\format{
An object of class \code{"tibble"}
}
\usage{
lasy.log.reg(
  independent.var,
  dependent.var,
  print.cov = FALSE,
  covariates = NULL,
  data
)
}
\arguments{
\item{independent.var}{independent variable/s}

\item{dependent.var}{dependent variable/s}

\item{print.cov}{Print effect of covariates, default is FALSE}

\item{covariates}{covariates to be included in a model}

\item{data}{data frame or tibble object}
}
\value{
data frame
}
\description{
This function performs logistic regression and print results in tibble output.
This function aims to provide the results of the regression analysis in the format, which is frequently
desired in academic journals.
}
\details{
Currently, this function does not provide model fit indicators such as AIC or BIC.
}
\examples{
# data loading
data(paq.validation.study)
regress.output <- lasy.log.reg(independent.var = c("TEQ","PAQ","G_EOT","G_DDF"),
                              covariates = c("Age"),
                              dependent.var = c("econom_stat_dich",
                                            "family_status",
                                            "edu_dich"),
                              data = paq.validation.study)

print(regress.output)
}
\references{
Welch, B. L. (1947). "The generalization of "Student's"
problem when several different population variances are involved".
Biometrika. 34 (1--2): 28--35.

Wilcoxon, F., Individual Comparisons by Ranking Methods,
Biometrics Bulletin, Vol. 1, 1945, pp. 80--83. DOI:10.2307/3001968

Dunn, O. J. (1961) Multiple comparisons among means.
Journal of the American Statistical Association. 56, 52--64.

Games, P. A., Keselman, H. J., & Clinch, J. J.
Tests for homogeneity of variance in factorial designs.
Psychological Bulletin, 86, 978--984
}
\author{
Lukas Novak, \email{lukasjirinovak@gmail.com}
}
\keyword{logistic-regression}
