% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitab.R
\docType{data}
\name{mitab}
\alias{mitab}
\title{Measurement invariance table}
\format{
An object of class \code{"tibble"}
}
\usage{
mitab(
  group1_nam,
  group2_nam,
  ordered,
  model,
  data,
  std.lv,
  meanstructure,
  group,
  yes_no_results,
  estimator,
  robust = FALSE,
  cfi.difference = FALSE,
  rmsea.difference = FALSE
)
}
\arguments{
\item{group1_nam}{name of the first group}

\item{group2_nam}{name of the second group}

\item{ordered}{logical, if set to TRUE items will be treated as ordered variables}

\item{model}{lavaan model to test}

\item{data}{data frame or tibble}

\item{std.lv}{logical, if TRUE than standardized loadings are stored in temporal output}

\item{meanstructure}{logical, if TRUE than model with meanstructure is estimated}

\item{group}{name of grouping variable}

\item{yes_no_results}{logical, if TRUE than lasy output indicating difference between models is added, currently working only based on CFI and RMSEA}

\item{estimator}{name of estimator to be used during fitting procedure}

\item{robust}{logical, if TRUE, than robust results are printed, working only with estimators  providing robust results (e.g. MLR or DWLS)}

\item{cfi.difference}{logical, if TRUE, delta of the CFI is printed in output}

\item{rmsea.difference}{logical, if TRUE, delta of RMSEA is printed in output}
}
\value{
data frame
}
\description{
Measurement invariance table
}
\details{
This function creates table with the key output from measurement invariance testing.
}
\examples{
# The famous Holzinger and Swineford (1939) example
HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '

library(lavaan)
dat <- HolzingerSwineford1939
res.tab.mi <- mitab(
group1_nam = "Grant-White",
group2_nam = "Pasteur",
ordered = FALSE,
model = HS.model,
data = dat,
std.lv = TRUE,
meanstructure = TRUE,
group = "school",
yes_no_results = TRUE,
estimator = "MLR",
robust = TRUE,
cfi.difference = TRUE,
rmsea.difference = TRUE
)

print(res.tab.mi)
}
\references{
Myles Hollander and Douglas A. Wolfe (1973). Nonparametric Statistical Methods. New York: John Wiley & Sons. Pages 27--33 (one-sample), 68--75 (two-sample).
Or second edition (1999).
}
\author{
Lukas Novak, \email{lukasjirinovak@gmail.com}
}
\keyword{MI,}
\keyword{a}
\keyword{equivalence,}
\keyword{invariance}
\keyword{measurement}
\keyword{of}
