% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.g.comp.R
\docType{data}
\name{mult.g.comp}
\alias{mult.g.comp}
\title{Automatic multi-group comparison}
\format{
An object of class \code{"tibble"}
}
\usage{
mult.g.comp(df, outcome.var, groups)
}
\arguments{
\item{df}{data frame or tibble with groups to compare}

\item{outcome.var}{continuous variables}

\item{groups}{grouping variable}
}
\value{
data frame
}
\description{
Automatic multi-group comparison
}
\details{
This function computes either Wilcox test or t-test depending on whether homogeneity of variances assumption is met or not.
}
\examples{
# data loading
data(dat)
# running the function
two.g.comp.out.EC = two.g.comp(df = dat, y = "IRI_EC", group.var = "Gender")
# printing the output
print(two.g.comp.out.EC)
}
\references{
Myles Hollander and Douglas A. Wolfe (1973). Nonparametric Statistical Methods. New York: John Wiley & Sons. Pages 27--33 (one-sample), 68--75 (two-sample).
Or second edition (1999).
}
\author{
Lukas Novak, \email{lukasjirinovak@gmail.com}
}
\keyword{Wilcoxon}
\keyword{comparison,}
\keyword{datasets,}
\keyword{group}
\keyword{test}
\keyword{two}
