% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychtoolbox-package.R
\docType{package}
\name{psychtoolbox-package}
\alias{psychtoolbox}
\alias{psychtoolbox-package}
\title{psychtoolbox: Tools for psychology research and psychometrics}
\description{
This package contains functions helping to analyse psychological data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/lukas.novak/psychtoolbox}
}

}
\author{
\strong{Maintainer}: Lukas Novak \email{lukasjirinovak@gmail.com} (\href{https://orcid.org/0000-0002-7582-2098}{ORCID})

}
\keyword{internal}
