% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCI.R
\docType{data}
\name{RCI}
\alias{RCI}
\title{Reliable Change Index (RCI)}
\format{
numeric vector of values
}
\usage{
RCI(SD_0, test.ret.rel)
}
\arguments{
\item{SD_0}{standard deviation of the non-clinical population}

\item{test.ret.rel}{test-retest reliability of the instrument}
}
\value{
numeric vector
}
\description{
This function calculates Reliable Change Index (RCI) as modifed by Wiger and Solberg (2001, p.148).
}
\details{
This function computes value corresponding to "the minimum amount of change that could not be attributed to the error of measurement" (Biescad & Timulak, 2014, p. 150). If score change from before to post treatment
is lower that value resulting from this function, than change in client score can not be attributed to the effectiveness of the therapy but rather other factors such as a measurement error (Biescad & Timulak, 2014). This function
is a result of modification of the original Jacobson and Truax (1991) formula by Wiger and Solberg (2001, p.148).
}
\examples{

re.ch.in = RCI(SD_0 = 4.87, test.ret.rel = 0.66)
re.ch.in

}
\references{
Jacobson, N. S., & Truax, P. (1991). Clinical significance: A statistical approach to defining meaningful change in psychotherapy research. Journal of Consulting and Clinical Psychology, 59(1), 12-19, DOI:
\url{https://doi.org/10.1037/0022-006X.59.1.12}

Matus Biescad & Ladislav Timulak (2014). Measuring psychotherapy
outcomes in routine practice: Examining Slovak versions of three commonly used outcome
instruments, European Journal of Psychotherapy & Counselling, 16:2, 140-162, DOI:
\url{https://doi.org/10.1080/13642537.2014.895772}

Wiger, D. E., & Solberg, K. B. (2001). Tracking Mental Health Outcomes:
A Therapist’s Guide to Measuring Client Progress, Analyzing Data,
and Improving Your Practice (1., Vol. 2001). Wiley.
}
\seealso{
\code{\link[=clin_sig_chang]{clin_sig_chang()}} function for calculation of the clinical cut-off scores
}
\author{
Lukas Novak, \email{lukasjirinovak@gmail.com}
}
