% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_boot_table.R
\docType{methods}
\name{small_boot_table}
\alias{small_boot_table}
\title{Summarizing Bootstrapped Network Estimates from 'bootnet'}
\format{
An object of class \code{"tibble"}.
}
\usage{
small_boot_table(bootnet_output, predict_function_output = NULL)
}
\arguments{
\item{bootnet_output}{output from the 'bootnet' package after bootstrapping network analysis.}

\item{predict_function_output}{optional, output from a predictability analysis function.}
}
\value{
A tibble summarizing edge weights and, optionally, node predictability.
}
\description{
Summarizing Bootstrapped Network Estimates from 'bootnet'
}
\details{
This function provides a convenient summary of bootstrapped network estimates from the 'bootnet' package, including edge weights and their confidence intervals. Optionally, it can also include node predictability metrics if provided. It is tailored to enhance the usability of 'bootnet' output by summarizing critical metrics in a concise format.
}
\examples{
  # Load required package
  library(bootnet)

  # Simulate data using base R
  set.seed(123)
  Sigma <- matrix(0.5, nrow = 5, ncol = 5) + diag(0.5, 5)
  chol_decomp <- chol(Sigma)
  z <- matrix(rnorm(100 * 5), nrow = 100, ncol = 5)
  sim_data <- z \%*\% chol_decomp
  sim_data_df <- as.data.frame(sim_data)

  # Estimate a network using 'EBICglasso'
  estimate <- bootnet::estimateNetwork(sim_data_df, default = "EBICglasso")

  # Perform bootstrapping on the estimated network
  boot_results <- bootnet::bootnet(estimate, nBoots = 100, nCores = 1)

  # Summarize the bootstrapped network
  summary <- small_boot_table(bootnet_output = boot_results)
  print(summary)
}
\references{
Epskamp, S., Borsboom, D., & Fried, E. I. (2018). Network analysis: An integrative approach to the structure of psychopathology. Annual review of clinical psychology, 14, 91-121.
}
\author{
Lukas Novak, \email{lukasjirinovak@gmail.com}
}
\keyword{'bootnet'}
\keyword{analysis,}
\keyword{bootstrapping,}
\keyword{network}
