% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_word.R
\name{word2pdf}
\alias{word2pdf}
\title{word to pdf}
\format{
An object of class \code{"pdf"}
}
\usage{
word2pdf(imp_file, out_file)
}
\arguments{
\item{imp_file}{name of the word document to convert - without docx suffix}

\item{out_file}{name of output pdf file without - without pdf suffix}
}
\value{
pdf file
}
\description{
Conversion of word document to pdf using either R Markdown package or Libre office. The latter represents higher quality approach - in general.
}
\details{
this function is currently running only on windows
}
\examples{
# example from word do pdf
#word2pdf(imp_file = "example.docx",out_file = "example1.pdf")
}
\author{
Lukas Novak, \email{lukasjirinovak@gmail.com}
}
\keyword{pdf,word}
