% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{dat}
\alias{dat}
\title{DATASET_TITLE}
\format{
A data frame with 835 rows and 2 variables:
\describe{
\item{\code{Gender}}{integer COLUMN_DESCRIPTION}
\item{\code{IRI_EC}}{double COLUMN_DESCRIPTION}
}
}
\usage{
dat
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
