% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitab.R
\name{mitab}
\alias{mitab}
\title{Measurement invariance table}
\usage{
mitab(
  group1_nam,
  group2_nam,
  ordered,
  model,
  data,
  std.lv,
  meanstructure,
  group,
  yes_no_results,
  estimator,
  robust = FALSE,
  cfi.difference = TRUE,
  rmsea.difference = TRUE,
  per_group_models = FALSE,
  ...
)
}
\arguments{
\item{group1_nam}{name of the first group. Used if per_group_models is TRUE.}

\item{group2_nam}{name of the second group. Used if per_group_models is TRUE.}

\item{ordered}{logical, if set to TRUE items will be treated as ordered variables}

\item{model}{lavaan model to test}

\item{data}{data frame or tibble}

\item{std.lv}{logical, if TRUE, the latent variables are standardized.}

\item{meanstructure}{logical, if TRUE, the model includes mean structures.}

\item{group}{name of grouping variable}

\item{yes_no_results}{logical, if TRUE, a column indicating significant model differences is added, based on CFI and RMSEA criteria.}

\item{estimator}{name of estimator to be used during fitting procedure (e.g., "MLR", "WLSMV").}

\item{robust}{logical, if TRUE, robust fit statistics are extracted. This is recommended for estimators like MLR or WLSMV.}

\item{cfi.difference}{logical, if TRUE, the change in CFI between nested models is included in the output.}

\item{rmsea.difference}{logical, if TRUE, the change in RMSEA between nested models is included in the output.}

\item{per_group_models}{logical, if TRUE, single-group models are fitted for each group and included in the output table. Defaults to FALSE.}

\item{...}{optional arguments to be passed to the lavaan::cfa function.}
}
\value{
A tibble containing a formatted summary of the measurement invariance testing results.
}
\description{
Measurement invariance table
}
\details{
This function automates measurement invariance testing using lavaan and presents the results in a clear, formatted table. It uses standard lavaan functions for all model fitting.
}
\examples{
# The famous Holzinger and Swineford (1939) example
HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '

library(lavaan)
dat <- HolzingerSwineford1939
# Example with a robust ML estimator (MLR)
res.tab.mlr <- mitab(
  group1_nam = "Grant-White",
  group2_nam = "Pasteur",
  ordered = FALSE,
  model = HS.model,
  data = dat,
  std.lv = TRUE,
  meanstructure = TRUE,
  group = "school",
  yes_no_results = TRUE,
  estimator = "MLR",
  robust = TRUE,
  cfi.difference = TRUE,
  rmsea.difference = TRUE,
  per_group_models = TRUE # Optionally show per-group fits
)

print(res.tab.mlr)

# Example with an ordered dataset and the WLSMV estimator
set.seed(123)
ord_data <- HolzingerSwineford1939
ord_data[, paste0("x", 1:9)] <- lapply(ord_data[, paste0("x", 1:9)],
                                       function(z) as.integer(cut(z, breaks = 4)))

res.tab.wlsmy <- mitab(
  group1_nam = "Grant-White",
  group2_nam = "Pasteur",
  ordered = TRUE,
  model = HS.model,
  data = ord_data,
  std.lv = TRUE,
  meanstructure = TRUE,
  group = "school",
  yes_no_results = TRUE,
  estimator = "WLSMV",
  robust = TRUE, # Using robust=TRUE is recommended for WLSMV
  cfi.difference = TRUE,
  rmsea.difference = TRUE
)
print(res.tab.wlsmy)
}
\references{
Cheung, G. W., & Rensvold, R. B. (2002). Evaluating goodness-of-fit indexes for testing measurement invariance. Structural Equation Modeling, 9(2), 233-255.
}
\author{
Lukas Novak, \email{lukasjirinovak@gmail.com}
}
\keyword{MI,}
\keyword{a}
\keyword{equivalence,}
\keyword{invariance}
\keyword{measurement}
\keyword{of}
