% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.g.comp.R
\docType{data}
\name{mult.g.comp}
\alias{mult.g.comp}
\title{Robust multi-group comparison}
\format{
An object of class \code{"tibble"}
}
\usage{
mult.g.comp(df, outcome.var, groups)
}
\arguments{
\item{df}{data frame or tibble object}

\item{outcome.var}{continuous variable/s}

\item{groups}{grouping variable/s}
}
\value{
data frame
}
\description{
This function allows to compare multiple groups in
multiple outcome variables with violated parametric assumptions.
}
\details{
Currently, this function does not report effect size from post-hoc tests.
\subsection{Two group comparison}{

If there is less than three groups, the Welch test or the Wilcoxon
test depending on data distribution.
}

\subsection{Three and more groups comparison}{

If more than two groups are present in data, the Dunn test or Games-Howell
test is performed.
}
}
\examples{
# data loading
data(dat)
tab.1=mult.g.comp(df = paq.validation.study, outcome.var = c("PAQ","G_DIF","G_DDF","G_EOT"),
groups = c("economical_status",
          "Gender",
          "education",
          "family_status"))
# printing the output
print(tab.1)
}
\references{
Welch, B. L. (1947). "The generalization of "Student's"
problem when several different population variances are involved".
Biometrika. 34 (1--2): 28--35.

Wilcoxon, F., Individual Comparisons by Ranking Methods,
Biometrics Bulletin, Vol. 1, 1945, pp. 80--83. DOI:10.2307/3001968

Dunn, O. J. (1961) Multiple comparisons among means.
Journal of the American Statistical Association. 56, 52--64.

Games, P. A., Keselman, H. J., & Clinch, J. J.
Tests for homogeneity of variance in factorial designs.
Psychological Bulletin, 86, 978--984
}
\author{
Lukas Novak, \email{lukasjirinovak@gmail.com}
}
\keyword{Dunn-test}
\keyword{Games-Howell}
\keyword{multiple-groups}
\keyword{test,}
\keyword{testing,}
