% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_boot_table.R
\docType{methods}
\name{small_boot_table}
\alias{small_boot_table}
\title{Summarizing Bootstrapped Network Estimates from 'bootnet'}
\format{
An object of class \code{"tibble"}.
}
\usage{
small_boot_table(
  bootnet_output,
  predict_function_output = NULL,
  include_sample_edge_weight = TRUE,
  include_p_values = "stars"
)
}
\arguments{
\item{bootnet_output}{output from the 'bootnet' package after bootstrapping network analysis.}

\item{predict_function_output}{optional, output from a predictability analysis function.}

\item{include_sample_edge_weight}{Logical, whether to include sample edge weight in the summary table.}

\item{include_p_values}{String, whether to include p-values as significance stars, add p-value column, or exclude p-values entirely. Options are "stars", "exact", or "none". Default is "none".}
}
\value{
A tibble summarizing edge weights and, optionally, node predictability.
}
\description{
Summarizing Bootstrapped Network Estimates from 'bootnet'
}
\details{
This function provides a convenient summary of bootstrapped network estimates from the 'bootnet' package, including edge weights, their confidence intervals, and optionally p-values and node predictability metrics if provided. It is tailored to enhance the usability of 'bootnet' output by summarizing critical metrics in a concise format. Importantly, it only summarizes results in edges that were non-zero on sample level. Edges that had zero edge weight in a sample are filtered out.

The empirical p-values are calculated using the following formula:
\deqn{p = \frac{\sum_{i=1}^{n} \mathbf{1}\{|x_i| \geq |x_0|\}}{n}}
where \eqn{x_0} is the original estimate and \eqn{x_i} are the bootstrap estimates.
}
\examples{
  # Load required package

  # Simulate data using base R
  set.seed(654899)
  Sigma <- matrix(0.5, nrow = 5, ncol = 5) + diag(0.5, 5)
  chol_decomp <- chol(Sigma)
  z <- matrix(rnorm(100 * 5), nrow = 100, ncol = 5)
  sim_data <- z \%*\% chol_decomp
  sim_data_df <- as.data.frame(sim_data)

  # Estimate a network using 'EBICglasso'
  estimate <- bootnet::estimateNetwork(sim_data_df, default = "EBICglasso")

  # Perform bootstrapping on the estimated network
  boot_results <- bootnet::bootnet(estimate, nBoots = 100, nCores = 1)

  # Summarize the bootstrapped network
  summary <- small_boot_table(bootnet_output = boot_results)
  print(summary)
}
\references{
Epskamp, S., Borsboom, D., & Fried, E. I. (2018). Network analysis: An integrative approach to the structure of psychopathology. Annual review of clinical psychology, 14, 91-121.
}
\author{
Lukas Novak, \email{lukasjirinovak@gmail.com}
}
\keyword{'bootnet'}
\keyword{analysis,}
\keyword{bootstrapping,}
\keyword{network}
