% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\docType{data}
\name{two.g.comp}
\alias{two.g.comp}
\title{Automatic post-hoc testing}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
two.g.comp(df, y, var.nam)
}
\arguments{
\item{df}{frame with one socio-demographic variable and one continous}

\item{y}{varialbe (continous)}

\item{var.nam}{variable (Socio-demographic variable)}
}
\value{
data frame
}
\description{
Automatic post-hoc testing
}
\examples{
data(dat)
# data loading
# running the function
two.g.comp.out.EC = two.g.comp(df = dat, y = "IRI_EC", var.nam = "Gender")
# printing the output
print(two.g.comp.out.EC)
}
\keyword{datasets}
